refactor repo to match current template
