+++ pyproject.toml
@@ -5,7 +5,6 @@ description = "arcaflow plugin for stressng"
 authors = ["arcalot"]
 license = "Apache-2.0+GPL-2.0-only"
 readme = "README.md"
-packages = [{include = "stressng_plugin.py", from="."}]
 
 [tool.poetry.dependencies]
 python = "^3.9"
@@ -14,7 +13,6 @@ scipy = "^1.9.3"
 pyyaml = ">=5.4,<6.0"
 arcaflow-plugin-sdk = "^0.10.1"
 
-
 [build-system]
-requires = ["poetry-core"]
+requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
